requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements-txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1 ; platform_python_implementation == 'PyPy'
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6 ; sys_platform == 'win32'
filetype==1.2.0
firecrawl-py==1.11.1
frozenlist==1.5.0
gitdb==4.0.12
gitpython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
greenlet==3.1.1 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.18
langchain-core==0.3.35
langchain-google-genai==2.0.9
langchain-text-splitters==0.3.6
langsmith==0.3.8
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.26.0
nest-asyncio==1.6.0
numpy==1.26.2
orjson==3.10.15 ; platform_python_implementation != 'PyPy'
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22 ; platform_python_implementation == 'PyPy'
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.38
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0 ; sys_platform != 'darwin'
websockets==14.2
yarl==1.18.3
zstandard==0.23.0
